// frontend/src/api/enums.ts
export enum FanType {
  Part = 1,
  Aux = 2,
  Chamber = 3,
}

export enum TemperatureType {
  Nozzle = 'nozzle',
  Heatbed = 'heatbed',
  Chamber = 'chamber',
}

export enum LightType {
  Chamber = 'chamber_light',
}

export enum LightMode {
  On = 'on',
  Off = 'off',
}

export enum GcodeState {
  Idle = 'IDLE',
  Prepare = 'PREPARE',
  Running = 'RUNNING',
  Pause = 'PAUSE',
  Finish = 'FINISH',
  Failed = 'FAILED',
  Unknown = 'UNKNOWN',
}

export enum PrintSpeedLevel {
  Silent = 1,
  Standard = 2,
  Sport = 3,
  Ludicrous = 4,
}

export enum CURRENT_STAGE_IDS {
  PRINTING = 0,
  AUTO_BED_LEVELING = 1,
  HEATBED_PREHEATING = 2,
  SWEEPING_XY_MECH_MODE = 3,
  CHANGING_FILAMENT = 4,
  M400_PAUSE = 5,
  PAUSED_FILAMENT_RUNOUT = 6,
  HEATING_HOTEND = 7,
  CALIBRATING_EXTRUSION = 8,
  SCANNING_BED_SURFACE = 9,
  INSPECTING_FIRST_LAYER = 10,
  IDENTIFYING_BUILD_PLATE_TYPE = 11,
  CALIBRATING_MICRO_LIDAR = 12,
  HOMING_TOOLHEAD = 13,
  CLEANING_NOZZLE_TIP = 14,
  CHECKING_EXTRUDER_TEMPERATURE = 15,
  PAUSED_USER = 16,
  PAUSED_FRONT_COVER_FALLING = 17,
  CALIBRATING_MICRO_LIDAR_18 = 18,
  CALIBRATING_EXTRUSION_FLOW = 19,
  PAUSED_NOZZLE_TEMPERATURE_MALFUNCTION = 20,
  PAUSED_HEAT_BED_TEMPERATURE_MALFUNCTION = 21,
  FILAMENT_UNLOADING = 22,
  PAUSED_SKIPPED_STEP = 23,
  FILAMENT_LOADING = 24,
  CALIBRATING_MOTOR_NOISE = 25,
  PAUSED_AMS_LOST = 26,
  PAUSED_LOW_FAN_SPEED_HEAT_BREAK = 27,
  PAUSED_CHAMBER_TEMPERATURE_CONTROL_ERROR = 28,
  COOLING_CHAMBER = 29,
  PAUSED_USER_GCODE = 30,
  MOTOR_NOISE_SHOWOFF = 31,
  PAUSED_NOZZLE_FILAMENT_COVERED_DETECTED = 32,
  PAUSED_CUTTER_ERROR = 33,
  PAUSED_FIRST_LAYER_ERROR = 34,
  PAUSED_NOZZLE_CLOG = 35,
  CHECK_ABSOLUTE_ACCURACY_BEFORE_CALIBRATION = 36,
  ABSOLUTE_ACCURACY_CALIBRATION = 37,
  CHECK_ABSOLUTE_ACCURACY_AFTER_CALIBRATION = 38,
  CALIBRATE_NOZZLE_OFFSET = 39,
  BED_LEVEL_HIGH_TEMPERATURE = 40,
  CHECK_QUICK_RELEASE = 41,
  CHECK_DOOR_AND_COVER = 42,
  LASER_CALIBRATION = 43,
  CHECK_PLAFORM = 44,
  CHECK_BIRDEYE_CAMERA_POSITION = 45,
  CALIBRATE_BIRDEYE_CAMERA = 46,
  BED_LEVEL_PHASE_1 = 47,
  BED_LEVEL_PHASE_2 = 48,
  HEATING_CHAMBER = 49,
  HEATED_BEDCOOLING = 50,
  PRINT_CALIBRATION_LINES = 51,
  CHECK_MATERIAL = 52,
  CALIBRATING_LIVE_VIEW_CAMERA = 53,
  WAITING_FOR_HEATBED_TEMPERATURE = 54,
  CHECK_MATERIAL_POSITION = 55,
  CALIBRATING_CUTTER_MODEL_OFFSET = 56,
  MEASURING_SURFACE = 57,
  THERMAL_PRECONDITIONING = 58,
  HOMING_BLADE_HOLDER = 59,
  CALIBRATING_CAMERA_OFFSET = 60,
  CALIBRATING_BLADE_HOLDER_POSITION = 61,
  HOTEND_PICK_PLACE_TEST = 62,
  WAITING_CHAMBER_TEMPERATURE_EQUALIZE = 63,
  PREPARING_HOTEND = 64,
  CALIBRATING_DETECTION_NOZZLE_CLUMPING = 65,
  PURIFYING_CHAMBER_AIR = 66,
  PREPARING_AMS = 77,
  IDLE = -1,
  IDLE_255 = 255,
}

export enum GCODE_STATE_OPTIONS {
  Failed = 'failed',
  Finish = 'finish',
  Idle = 'idle',
  Init = 'init',
  Offline = 'offline',
  Pause = 'pause',
  Prepare = 'prepare',
  Running = 'running',
  Slicing = 'slicing',
  Unknown = 'unknown',
}